requirements.txt
speedtest-cli==2.1.2
mysql-connector-python==8.0.22
SQLAlchemy==1.3.20
mysql-python==1.2.5